+++ app/build.gradle
@@ -7,7 +7,7 @@ android {
     buildToolsVersion "30.0.3"
 
     defaultConfig {
-        applicationId "com.example.headi"
+        applicationId "com.headi.app"
         minSdkVersion 28
         targetSdkVersion 30
         versionCode 1
