+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.headi">
+    package="com.headi.app">
 
     <uses-permission android:name="android.permission.FOREGROUND_SERVICE" />
 
@@ -18,7 +18,7 @@
             android:resource="@xml/unused_to_delete" />
 
         <activity
-            android:name=".MainActivity"
+            android:name="com.headi.app.MainActivity"
             android:label="@string/app_name"
             android:theme="@style/Theme.Headi.NoActionBar">
             <intent-filter>
@@ -29,7 +29,7 @@
         </activity>
 
         <service
-            android:name=".TimerForegroundService"
+            android:name="com.headi.app.TimerForegroundService"
             android:enabled="true" />
 
     </application>
