--- app/src/main/java/com/example/headi/db/MedicationsCourserAdapter.java
@@ -1,40 +1,0 @@
-package com.example.headi.db;
-
-import android.content.Context;
-import android.database.Cursor;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.TextView;
-
-import androidx.cursoradapter.widget.CursorAdapter;
-
-import com.example.headi.R;
-
-public class MedicationsCourserAdapter extends CursorAdapter {
-
-    public MedicationsCourserAdapter(Context context, Cursor c, int flags) {
-        super(context, c, flags);
-    }
-
-    // The newView method is used to inflate a new view and return it,
-    // you don't bind any data to the view at this point.
-    @Override
-    public View newView(Context context, Cursor cursor, ViewGroup parent) {
-        return LayoutInflater.from(context).inflate(R.layout.fragment_pains_mediactions_item, parent, false);
-    }
-
-    // The bindView method is used to bind all data to a given view
-    // such as setting the text on a TextView.
-    @Override
-    public void bindView(View view, Context context, Cursor cursor) {
-        // Find fields to populate in inflated template
-        TextView pains_name = view.findViewById(R.id.pains_name);
-
-        // Extract properties from cursor
-        String pain = cursor.getString(cursor.getColumnIndexOrThrow(HeadiDBContract.Medication.COLUMN_MEDICATION));
-
-        // Populate fields with extracted properties
-        pains_name.setText(pain);
-    }
-}
